Added jwt extended and refresh token

+++ resources/item.py
@@ -1,63 +1,90 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import (
+    jwt_required, 
+    get_jwt, 
+    get_jwt_identity
+)
 from models.item import ItemModel
+
 
 class Item(Resource):
     parser = reqparse.RequestParser()
     parser.add_argument('price',
-        type=float,
-        required=True,
-        help="This field cannot be left blank"
-    )
+                        type=float,
+                        required=True,
+                        help="This field cannot be left blank!"
+                        )
     parser.add_argument('store_id',
-        type=int,
-        required=True,
-        help="Every item needs a store id"
-    )
+                        type=int,
+                        required=True,
+                        help="Every item needs a store_id."
+                        )
 
-    @jwt_required()
+    @jwt_required()  # No longer needs brackets
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
             return item.json()
-        return {"message": "Item not found"}, 404
+        return {'message': 'Item not found'}, 404
 
+    @jwt_required(fresh=True)
     def post(self, name):
         if ItemModel.find_by_name(name):
-            return {'message': "An item with name'{}' already exists.".format(name)}, 400
+            return {'message': "An item with name '{}' already exists.".format(name)}, 400
+
         data = Item.parser.parse_args()
-        item = ItemModel(name, data['price'], data['store_id'])
+
+        item = ItemModel(name, **data)
+
         try:
             item.save_to_db()
         except:
-            return {"message": "An error occured inserting the item"}, 500
-        
+            return {"message": "An error occurred inserting the item."}, 500
+
         return item.json(), 201
-    
+
+    @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
-        
-        return {'message': 'Item deleted'}
-    
+            return {'message': 'Item deleted.'}
+        return {'message': 'Item not found.'}, 404
+
     def put(self, name):
         data = Item.parser.parse_args()
+
         item = ItemModel.find_by_name(name)
-        updated_item = ItemModel(name, **data)
-        if item is None:
-            try:
-                item = ItemModel(name, **data)
-            except:
-                return {"message": "An error occured inserting the item"}, 500
-        else:
+
+        if item:
             item.price = data['price']
-            iteö.store_id = data['store_id']
-        
+        else:
+            item = ItemModel(name, **data)
+
         item.save_to_db()
 
         return item.json()
 
+
 class ItemList(Resource):
+    @jwt_required(optional=True)
     def get(self):
-        return {'items': [item.json() for item in ItemModel.query.all()]}
+        """
+        Here we get the JWT identity, and then if the user is logged in (we were able to get an identity)
+        we return the entire item list.
+        Otherwise we just return the item names.
+        This could be done with e.g. see orders that have been placed, but not see details about the orders
+        unless the user has logged in.
+        """
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200
+        return {
+            'items': [item['name'] for item in items],
+            'message': 'More data available if you log in.'
+        }, 200
